+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 
 ### Breaking Changes
 
+* Driver version greater or equal to `v2.0.20250502dev`
+  * Changes to `config.default.ini`: `SOC_RESET_VOLTAGE` was replaced by `SOC_RESET_CELL_VOLTAGE`
+
 * Driver version greater or equal to `v2.0.20250207dev`
   * Changes to `config.default.ini`: `TEMPERATURE_SOURCE_BATTERY` is now a list of temperature sensors, so you can choose which sensors you want to use
 
@@ -69,7 +72,7 @@
 * Added: Signal handler for clean service restart/shutdown by @mr-manuel
 * Added: UBMS CAN - support for Valence U-BMS by @gimx
 * Changed: A lot of under the hood optimizations by @mr-manuel
-* Changed: Apply `SOC_RESET_VOLTAGE` after `SOC_RESET_AFTER_DAYS` regardless of whether the battery is in absorption, bulk, or float mode https://github.com/mr-manuel/venus-os_dbus-serialbattery/issues/123 by @mr-manuel
+* Changed: Apply `SOC_RESET_CELL_VOLTAGE` after `SOC_RESET_AFTER_DAYS` regardless of whether the battery is in absorption, bulk, or float mode https://github.com/mr-manuel/venus-os_dbus-serialbattery/issues/123 by @mr-manuel
 * Changed: BLE - Reworked log notifications be more helpful by @mr-manuel
 * Changed: BLE - Reworked installation for external Bluetooth USB dongle by @Felixrising & @mr-manuel
 * Changed: Consumed capacity must be negative values by @mr-manuel
@@ -113,6 +116,7 @@
 * Changed: The setting `SOC_CALC_CURRENT_MEASURED_BY_USER` was replaced by `CURRENT_MEASURED_BY_USER` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `SOC_CALC_CURRENT_REPORTED_BY_BMS` was replaced by `CURRENT_REPORTED_BY_BMS` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `SOC_LEVEL_TO_RESET_VOLTAGE_LIMIT` was replaced by `SWITCH_TO_BULK_SOC_THRESHOLD` in the `config.default.ini` by @mr-manuel
+* Changed: The setting `SOC_RESET_VOLTAGE` was replaced by `SOC_RESET_CELL_VOLTAGE` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `TEMP_1_NAME` was replaced by `TEMPERATURE_1_NAME` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `TEMP_2_NAME` was replaced by `TEMPERATURE_2_NAME` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `TEMP_3_NAME` was replaced by `TEMPERATURE_3_NAME` in the `config.default.ini` by @mr-manuel